DPRO-2657: Add bucket name to identifiers; support read services
Pulled the bucket-name+key pair into the new RepoId class and compose it
into the other identifier classes.

The default bucket name has been removed from ContentRepoServiceImpl, and
it has been updated to use the bucket name from all of those updated
identifier classes. It still needs to get a bucket name for destructive
operations, which will require further changes to the input classes. It is
in a non-compiling state for this reason.